refactor(storagebackend): move engine config and setup out of cmd/oteldb
The embedded engine's config surface and construction lived in package main, so only oteldb
could open one. Moves them to storagebackend as Config/Open, leaving the YAML shape and every
field name unchanged, so a second binary can open the same engine from the same config.

Config-parsing tests stay in cmd/oteldb; option-translation tests move with the code.
