+++ internal/storagebackend/open_internal_test.go
@@ -1,4 +1,4 @@
-package main
+package storagebackend
 
 import (
 	"context"
@@ -35,13 +35,13 @@ func TestClusterOption(t *testing.T) {
 		require.NoError(t, err)
 		require.Nil(t, opt)
 
-		opt, err = clusterOption(&StorageClusterConfig{ID: "n1"}, lg) // no etcd endpoints
+		opt, err = clusterOption(&ClusterConfig{ID: "n1"}, lg) // no etcd endpoints
 		require.NoError(t, err)
 		require.Nil(t, opt)
 	})
 
 	t.Run("Explicit", func(t *testing.T) {
-		opt, err := clusterOption(&StorageClusterConfig{
+		opt, err := clusterOption(&ClusterConfig{
 			Etcd: []string{"http://etcd:2379"}, ID: "node-1", Zone: "z1",
 			Addr: "node-1:9000", RF: 2, ShardsPerTenant: 4, Root: "/x",
 		}, lg)
@@ -58,12 +58,12 @@ func TestClusterOption(t *testing.T) {
 	})
 
 	t.Run("AddrDerivedFromID", func(t *testing.T) {
-		opt, err := clusterOption(&StorageClusterConfig{Etcd: []string{"e"}, ID: "node-7"}, lg)
+		opt, err := clusterOption(&ClusterConfig{Etcd: []string{"e"}, ID: "node-7"}, lg)
 		require.NoError(t, err)
 		o := applyOption(t, opt)
 		require.Equal(t, "node-7:7946", o.Cluster.Self.Addr, "default replication port")
 
-		opt, err = clusterOption(&StorageClusterConfig{Etcd: []string{"e"}, ID: "node-7", Port: 1234}, lg)
+		opt, err = clusterOption(&ClusterConfig{Etcd: []string{"e"}, ID: "node-7", Port: 1234}, lg)
 		require.NoError(t, err)
 		o = applyOption(t, opt)
 		require.Equal(t, "node-7:1234", o.Cluster.Self.Addr)
@@ -72,49 +72,12 @@ func TestClusterOption(t *testing.T) {
 	t.Run("IDAndAddrDefaultToHostname", func(t *testing.T) {
 		host, err := os.Hostname()
 		require.NoError(t, err)
-		opt, err := clusterOption(&StorageClusterConfig{Etcd: []string{"e"}}, lg)
+		opt, err := clusterOption(&ClusterConfig{Etcd: []string{"e"}}, lg)
 		require.NoError(t, err)
 		o := applyOption(t, opt)
 		require.Equal(t, host, o.Cluster.Self.ID)
 		require.Equal(t, host+":7946", o.Cluster.Self.Addr)
 	})
-}
-
-// TestStorageClusterConfigYAML checks the cluster block parses from the documented YAML shape via
-// the real config loader.
-func TestStorageClusterConfigYAML(t *testing.T) {
-	const data = `
-storage:
-  backend: file
-  dir: /data
-  cluster:
-    etcd: ["http://etcd-1:2379", "http://etcd-2:2379"]
-    id: oteldb-1
-    zone: a
-    addr: oteldb-1:7946
-    rf: 3
-    shards_per_tenant: 8
-    root: /oteldb
-`
-	f, err := os.CreateTemp("", "oteldb.yml")
-	require.NoError(t, err)
-	defer func() { _ = os.Remove(f.Name()) }()
-	_, err = f.WriteString(data)
-	require.NoError(t, err)
-	require.NoError(t, f.Close())
-
-	cfg, err := loadConfig(f.Name())
-	require.NoError(t, err)
-
-	require.Equal(t, "file", cfg.Storage.Backend)
-	require.NotNil(t, cfg.Storage.Cluster)
-	require.Equal(t, []string{"http://etcd-1:2379", "http://etcd-2:2379"}, cfg.Storage.Cluster.Etcd)
-	require.Equal(t, "oteldb-1", cfg.Storage.Cluster.ID)
-	require.Equal(t, "a", cfg.Storage.Cluster.Zone)
-	require.Equal(t, "oteldb-1:7946", cfg.Storage.Cluster.Addr)
-	require.Equal(t, 3, cfg.Storage.Cluster.RF)
-	require.Equal(t, 8, cfg.Storage.Cluster.ShardsPerTenant)
-	require.Equal(t, "/oteldb", cfg.Storage.Cluster.Root)
 }
 
 func TestTenancyOption(t *testing.T) {
@@ -123,13 +86,13 @@ func TestTenancyOption(t *testing.T) {
 		require.NoError(t, err)
 		require.Nil(t, opt)
 
-		opt, err = tenancyOption(&StoragePolicyConfig{}) // no tiers, no recompress
+		opt, err = tenancyOption(&PolicyConfig{}) // no tiers, no recompress
 		require.NoError(t, err)
 		require.Nil(t, opt)
 	})
 
 	t.Run("ResolvesPolicyForEveryTenant", func(t *testing.T) {
-		opt, err := tenancyOption(&StoragePolicyConfig{
+		opt, err := tenancyOption(&PolicyConfig{
 			Precision: []PrecisionTierConfig{
 				{After: 7 * 24 * time.Hour, Bits: 32},
 				{After: 30 * 24 * time.Hour, Bits: 16},
@@ -183,7 +146,7 @@ func TestTenancyOption(t *testing.T) {
 	})
 
 	t.Run("RetentionOnlyInstallsResolver", func(t *testing.T) {
-		opt, err := tenancyOption(&StoragePolicyConfig{
+		opt, err := tenancyOption(&PolicyConfig{
 			Retention: &RetentionConfig{MaxAge: 14 * 24 * time.Hour},
 		})
 		require.NoError(t, err)
@@ -193,7 +156,7 @@ func TestTenancyOption(t *testing.T) {
 	})
 
 	t.Run("LimitsOnlyInstallsResolver", func(t *testing.T) {
-		opt, err := tenancyOption(&StoragePolicyConfig{
+		opt, err := tenancyOption(&PolicyConfig{
 			Limits: &LimitsConfig{MaxSeries: 1000},
 		})
 		require.NoError(t, err)
@@ -203,81 +166,25 @@ func TestTenancyOption(t *testing.T) {
 	})
 
 	t.Run("UnknownAggIsAnError", func(t *testing.T) {
-		_, err := tenancyOption(&StoragePolicyConfig{
+		_, err := tenancyOption(&PolicyConfig{
 			Downsample: []DownsampleTierConfig{{Interval: time.Minute, Agg: "median"}},
 		})
 		require.Error(t, err)
 	})
 
 	t.Run("NegativeRetentionIsAnError", func(t *testing.T) {
-		_, err := tenancyOption(&StoragePolicyConfig{
+		_, err := tenancyOption(&PolicyConfig{
 			Retention: &RetentionConfig{MaxAge: -time.Hour},
 		})
 		require.Error(t, err)
 	})
 
 	t.Run("SoftSeriesAboveHardIsAnError", func(t *testing.T) {
-		_, err := tenancyOption(&StoragePolicyConfig{
+		_, err := tenancyOption(&PolicyConfig{
 			Limits: &LimitsConfig{MaxSeries: 100, MaxSeriesSoft: 200},
 		})
 		require.Error(t, err)
 	})
-}
-
-// TestStoragePolicyConfigYAML checks the policy block parses from the documented YAML shape and
-// resolves into the storage tenancy policy.
-func TestStoragePolicyConfigYAML(t *testing.T) {
-	const data = `
-storage:
-  backend: file
-  dir: /data
-  policy:
-    precision:
-      - after: 168h
-        bits: 32
-    downsample:
-      - after: 24h
-        interval: 5m
-        agg: avg
-    recompress:
-      after: 336h
-      level: 9
-    retention:
-      max_age: 720h
-      max_bytes: 100GB
-    limits:
-      ingest_bytes_per_second: 10MB
-      max_series: 1000000
-      max_series_soft: 800000
-`
-	f, err := os.CreateTemp("", "oteldb.yml")
-	require.NoError(t, err)
-	defer func() { _ = os.Remove(f.Name()) }()
-	_, err = f.WriteString(data)
-	require.NoError(t, err)
-	require.NoError(t, f.Close())
-
-	cfg, err := loadConfig(f.Name())
-	require.NoError(t, err)
-
-	require.NotNil(t, cfg.Storage.Policy)
-	opt, err := tenancyOption(cfg.Storage.Policy)
-	require.NoError(t, err)
-	o := applyOption(t, opt)
-
-	p := o.Tenancy.Resolve("default")
-	require.Len(t, p.Precision.Tiers, 1)
-	require.Equal(t, 168*time.Hour, p.Precision.Tiers[0].After)
-	require.Equal(t, uint8(32), p.Precision.Tiers[0].Bits)
-	require.Len(t, p.Downsample.Tiers, 1)
-	require.Equal(t, signal.AggAvg, p.Downsample.Tiers[0].Agg)
-	require.Equal(t, 336*time.Hour, p.Recompress.After)
-	require.Equal(t, 9, p.Recompress.Level)
-	require.Equal(t, 720*time.Hour, p.Retention.MaxAge)
-	require.Equal(t, int64(100_000_000_000), p.Retention.MaxBytes)
-	require.Equal(t, int64(10_000_000), p.Limits.IngestBytesPerSecond)
-	require.Equal(t, int64(1_000_000), p.Limits.MaxSeries)
-	require.Equal(t, int64(800_000), p.Limits.MaxSeriesSoft)
 }
 
 func TestS3Backend(t *testing.T) {
@@ -287,12 +194,12 @@ func TestS3Backend(t *testing.T) {
 		_, err := s3Backend(ctx, nil)
 		require.Error(t, err)
 
-		_, err = s3Backend(ctx, &StorageS3Config{Region: "us-east-1"}) // no bucket
+		_, err = s3Backend(ctx, &S3Config{Region: "us-east-1"}) // no bucket
 		require.Error(t, err)
 	})
 
 	t.Run("StaticCredentials", func(t *testing.T) {
-		b, err := s3Backend(ctx, &StorageS3Config{
+		b, err := s3Backend(ctx, &S3Config{
 			Bucket:          "oteldb",
 			Prefix:          "data/",
 			Region:          "us-east-1",
@@ -307,64 +214,22 @@ func TestS3Backend(t *testing.T) {
 	})
 
 	t.Run("UnknownRetryProfile", func(t *testing.T) {
-		_, err := s3Backend(ctx, &StorageS3Config{Bucket: "oteldb", AccessKeyID: "k", SecretAccessKey: "s", Retry: "bogus"})
+		_, err := s3Backend(ctx, &S3Config{Bucket: "oteldb", AccessKeyID: "k", SecretAccessKey: "s", Retry: "bogus"})
 		require.Error(t, err)
 	})
 
 	t.Run("RetryProfiles", func(t *testing.T) {
 		for _, profile := range []string{"", "none", "default", "lossy"} {
-			b, err := s3Backend(ctx, &StorageS3Config{Bucket: "oteldb", AccessKeyID: "k", SecretAccessKey: "s", Retry: profile})
+			b, err := s3Backend(ctx, &S3Config{Bucket: "oteldb", AccessKeyID: "k", SecretAccessKey: "s", Retry: profile})
 			require.NoError(t, err, "profile %q", profile)
 			require.NotNil(t, b)
 		}
 	})
 }
 
-// TestStorageS3ConfigYAML checks the s3 block parses from the documented YAML shape via the real
-// config loader.
-func TestStorageS3ConfigYAML(t *testing.T) {
-	const data = `
-storage:
-  backend: s3
-  wal_dir: /var/lib/oteldb/wal
-  s3:
-    bucket: oteldb
-    prefix: data/
-    region: us-east-1
-    endpoint: http://minio:9000
-    force_path_style: true
-    access_key_id: key
-    secret_access_key: secret
-    session_token: token
-    retry: lossy
-`
-	f, err := os.CreateTemp("", "oteldb.yml")
-	require.NoError(t, err)
-	defer func() { _ = os.Remove(f.Name()) }()
-	_, err = f.WriteString(data)
-	require.NoError(t, err)
-	require.NoError(t, f.Close())
-
-	cfg, err := loadConfig(f.Name())
-	require.NoError(t, err)
-
-	require.Equal(t, "s3", cfg.Storage.Backend)
-	require.Equal(t, "/var/lib/oteldb/wal", cfg.Storage.WALDir)
-	require.NotNil(t, cfg.Storage.S3)
-	require.Equal(t, "oteldb", cfg.Storage.S3.Bucket)
-	require.Equal(t, "data/", cfg.Storage.S3.Prefix)
-	require.Equal(t, "us-east-1", cfg.Storage.S3.Region)
-	require.Equal(t, "http://minio:9000", cfg.Storage.S3.Endpoint)
-	require.True(t, cfg.Storage.S3.ForcePathStyle)
-	require.Equal(t, "key", cfg.Storage.S3.AccessKeyID)
-	require.Equal(t, "secret", cfg.Storage.S3.SecretAccessKey)
-	require.Equal(t, "token", cfg.Storage.S3.SessionToken)
-	require.Equal(t, "lossy", cfg.Storage.S3.Retry)
-}
-
 func TestResolveCacheSettings(t *testing.T) {
 	t.Run("DefaultsEnableAll", func(t *testing.T) {
-		s := resolveCacheSettings(StorageConfig{})
+		s := resolveCacheSettings(Config{})
 		// All four are opt-out for oteldb: on by default.
 		require.True(t, s.AggregateStats)
 		require.Greater(t, s.ReadCache, int64(0))
@@ -376,17 +241,17 @@ func TestResolveCacheSettings(t *testing.T) {
 
 	t.Run("MergeMemoryPassesThrough", func(t *testing.T) {
 		mm := xbytes.Bytes(4 << 20)
-		s := resolveCacheSettings(StorageConfig{MergeMemoryBytes: &mm})
+		s := resolveCacheSettings(Config{MergeMemoryBytes: &mm})
 		require.Equal(t, int64(4<<20), s.MergeMemory)
 
 		unbounded := xbytes.Bytes(-1)
-		s = resolveCacheSettings(StorageConfig{MergeMemoryBytes: &unbounded})
+		s = resolveCacheSettings(Config{MergeMemoryBytes: &unbounded})
 		require.Equal(t, int64(-1), s.MergeMemory, "negative means unbounded, not unset")
 	})
 
 	t.Run("ExplicitSizesHonored", func(t *testing.T) {
 		rc, dc, dm := xbytes.Bytes(1<<20), xbytes.Bytes(2<<20), xbytes.Bytes(3<<20)
-		s := resolveCacheSettings(StorageConfig{
+		s := resolveCacheSettings(Config{
 			ReadCacheBytes:    &rc,
 			DecodeCacheBytes:  &dc,
 			DecodeMemoryBytes: &dm,
@@ -398,14 +263,14 @@ func TestResolveCacheSettings(t *testing.T) {
 
 	t.Run("ZeroDisablesByteCache", func(t *testing.T) {
 		zero := xbytes.Bytes(0)
-		s := resolveCacheSettings(StorageConfig{ReadCacheBytes: &zero, DecodeCacheBytes: &zero, DecodeMemoryBytes: &zero})
+		s := resolveCacheSettings(Config{ReadCacheBytes: &zero, DecodeCacheBytes: &zero, DecodeMemoryBytes: &zero})
 		require.Equal(t, int64(0), s.ReadCache)
 		require.Equal(t, int64(0), s.DecodeCache)
 		require.Equal(t, int64(0), s.DecodeMemory, "explicit 0 disables decode admission control")
 	})
 
 	t.Run("AggregateStatsFalseDisables", func(t *testing.T) {
-		s := resolveCacheSettings(StorageConfig{AggregateStats: new(bool)})
+		s := resolveCacheSettings(Config{AggregateStats: new(bool)})
 		require.False(t, s.AggregateStats)
 	})
 }
@@ -507,34 +372,4 @@ func TestCacheOptions(t *testing.T) {
 		o := apply(t, cacheOptions(cacheSettings{AggregateStats: false}))
 		require.False(t, o.AggregateStats, "AggregateStats stays the library default when disabled")
 	})
-}
-
-// TestStorageCacheConfigYAML checks the cache block parses from the documented YAML shape, including
-// the opt-out semantics (an explicit 0 disables a byte cache, aggregate_stats: false disables the
-// sidecar).
-func TestStorageCacheConfigYAML(t *testing.T) {
-	const data = `
-storage:
-  backend: file
-  dir: /data
-  read_cache_bytes: "0"
-  decode_cache_bytes: "64MiB"
-  decode_memory_bytes: "256MiB"
-  aggregate_stats: false
-`
-	f, err := os.CreateTemp("", "oteldb.yml")
-	require.NoError(t, err)
-	defer func() { _ = os.Remove(f.Name()) }()
-	_, err = f.WriteString(data)
-	require.NoError(t, err)
-	require.NoError(t, f.Close())
-
-	cfg, err := loadConfig(f.Name())
-	require.NoError(t, err)
-
-	s := resolveCacheSettings(cfg.Storage)
-	require.Equal(t, int64(0), s.ReadCache, "explicit 0 disables the read cache")
-	require.Equal(t, int64(64<<20), s.DecodeCache)
-	require.Equal(t, int64(256<<20), s.DecodeMemory)
-	require.False(t, s.AggregateStats, "explicit aggregate_stats:false disables the sidecar")
 }
